Src/requirements.txt
imageio==2.28.1
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
setuptools==59.5.0
SimpleITK==2.2.1
torch==1.13.1
torch-ema==0.3
torchmetrics==0.10.3
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
matplotlib==3.5.3

